+++ medusa/server/web/home/handler.py
@@ -82,7 +82,6 @@ from medusa.search.manual import (
 from medusa.search.queue import (
     BacklogQueueItem,
     FailedQueueItem,
-    ForcedSearchQueueItem,
     SnatchQueueItem,
 )
 from medusa.server.web.core import (
@@ -1710,10 +1709,8 @@ class Home(WebRoot):
 
         return self.redirect('/home/displayShow?indexername={series_obj.indexer_name}&seriesid={series_obj.series_id}'.format(series_obj=series_obj))
 
-    def searchEpisode(self, indexername=None, seriesid=None, season=None, episode=None, manual_search=None):
-        """Search a ForcedSearch single episode using providers which are backlog enabled."""
-        down_cur_quality = 0
-
+    def searchEpisode(self, indexername=None, seriesid=None, season=None, episode=None):
+        """Search for a single episode using a Backlog Search using providers that are backlog enabled."""
         # retrieve the episode object and fail if we can't get one
         series_obj = Show.find_by_id(app.showList, indexer_name_to_id(indexername), seriesid)
         ep_obj = series_obj.get_episode(season, episode)
@@ -1723,7 +1720,7 @@ class Home(WebRoot):
             })
 
         # make a queue item for it and put it on the queue
-        ep_queue_item = ForcedSearchQueueItem(ep_obj.series, [ep_obj], bool(int(down_cur_quality)), bool(manual_search))
+        ep_queue_item = BacklogQueueItem(ep_obj.series, [ep_obj])
 
         app.forced_search_queue_scheduler.action.add_item(ep_queue_item)